Extracted serialization bits into own file.
